+++ js/consts.js
@@ -1,0 +1,41 @@
+const AVATAR_AMT = 6;
+const PICTURES_AMT = 25;
+const COMMENTS_AMT = 30;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const NAMES = [
+  'Дмитрий',
+  'Павел',
+  'Артём',
+  'Никита',
+  'Егор',
+  'Дарья',
+  'Ольга',
+  'Наталья'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Поймал момент, который хочется сохранить навсегда.',
+  'Солнце в душе и ветер в волосах.',
+  'Иногда нужно просто остановиться и посмотреть, как прекрасен мир.',
+  'Эту атмосферу не передать словами, только тишиной.',
+  'Маленькие радости — мой главный девиз по жизни.',
+  'Говорят, счастье в мелочах, и я с этим полностью согласен',
+  'Пятница, наконец-то! Планы на выходные?',
+  'Нашла свой кусочек счастья там, где его не ждала.',
+  'Это та самая магия, ради которой всё и затевалось.',
+  'Настроение: сиять изнутри.',
+  'Включила режим «наслаждаться моментом».',
+  'Просто я и мой кофе против всего мира.'
+];
+
+export {AVATAR_AMT, PICTURES_AMT, COMMENTS_AMT, MIN_LIKES, MAX_LIKES, NAMES, MESSAGES, DESCRIPTIONS};
